refactor: clean up the code before pull request
