+++ fs/moveFile.js
@@ -1,11 +1,11 @@
 import { parse, resolve } from 'path';
 import { createReadStream, createWriteStream } from 'fs';
 import { pipeline } from 'stream/promises';
-import { retrieveFilePaths } from '../init/retrieveFilePaths.js';
+import { retrieveFilePaths } from '../utils/helpers/retrieveFilePaths.js';
 import { stat, unlink } from 'fs/promises';
 import { getCurrentPath } from '../nwd/getCurrentPath.js';
 import { INVALID_INPUT_ERROR } from '../utils/constants/messages.js';
-import { getErrorMessage } from '../init/getErrorMessage.js';
+import { getErrorMessage } from '../utils/helpers/getErrorMessage.js';
 
 export const moveFile = async (paths) => {
 	try {
