+++ package.json
@@ -5,8 +5,7 @@
   "main": "index.js",
   "type": "module",
   "scripts": {
-    "start": "node index.js",
-    "test": "node index.js -- --username=Dasha"
+    "start": "node index.js"
   },
   "repository": {
     "type": "git",
