+++ utils/helpers/getErrorMessage.js
@@ -1,4 +1,4 @@
-import { INVALID_INPUT_ERROR, OPERATION_FAILED_ERROR } from '../utils/constants/messages.js';
+import { INVALID_INPUT_ERROR, OPERATION_FAILED_ERROR } from '../constants/messages.js';
 
 export const getErrorMessage = (message) => {
 	return message === INVALID_INPUT_ERROR
