+++ zip/decompressFile.js
@@ -2,11 +2,11 @@ import { basename, resolve } from 'path';
 import { createReadStream, createWriteStream } from 'fs';
 import { pipeline } from 'stream/promises';
 import { createBrotliDecompress } from 'zlib';
-import { retrieveFilePaths } from '../init/retrieveFilePaths.js';
+import { retrieveFilePaths } from '../utils/helpers/retrieveFilePaths.js';
 import { stat } from 'fs/promises';
 import { getCurrentPath } from '../nwd/getCurrentPath.js';
 import { INVALID_INPUT_ERROR } from '../utils/constants/messages.js';
-import { getErrorMessage } from '../init/getErrorMessage.js';
+import { getErrorMessage } from '../utils/helpers/getErrorMessage.js';
 
 export const decompressFile = async (paths) => {
 	try {
